Fixed YAML serialization of a stubbed object.
Previously, the mock proxy would be serialized when #to_yaml is called on a stubbed object.  The mock proxy includes a reference to a proc.  When the yaml string is deserialized with YAML.load, you'd get an "allocator undefined for Proc" TypeError.

This fix prevents the mock proxy from being serialized, so that the object is serialized the same as it would if it hadn't been stubbed.

Note that Marshaling also breaks when Marshal.dump is passed a stubbed object, with a "singleton can't be dumped" error.  I'm not yet sure how to fix this, so I've left a pending spec for it.