Implemented anounce_not_fulfilled_filters in the configuration.
It is called as last part of the CommandLine run initialization.
include_or_extend_modules was changed from an array to an hash.
Using the original Array values as keys and the symbols :was_fulfilled
and :not_fulfilled as values to mark whether a certain filter was
ever fulfilled. Unfortunately includes whith an empty hash as filters
seem to be never included - I'm not sure whether this is correct
or just an error in my not_fulfilled analysis.