Cleanup whitespace.
(Actually, I'm mostly doing this just to trigger a travis build. A
recent change in rspec-expectations[1] broke the rspec-mocks
travis build, and now that it's fixed[2], I want to see that the
rspec-mocks build is passing.  I figured I may as well make a useful
but non functional change while I'm at it, though).

[1] https://github.com/rspec/rspec-expectations/commit/40243447dd84aa5ec7cca13cf162f1f392d1bc29#commitcomment-1496381
[2] https://github.com/rspec/rspec-expectations/commit/a3e283979ab5f2e1a77eb5a14d55379a9aa7241d