Introdce config.set_proc_title
config.set_proc_title is a setting to enable a feature of example
to set the title of process of rspec to be the current example location.

If config.set_proc_title set to true, the process title is set to
the current example location and the full description of the example.

If config.set_proc_title set to `:simple`, the process title is set to
the current example location.