We only want to print out command if verbose is set
I've added this because I've noticed that on spec directories where there are a lot of files the header of the test and the footer of the test tend to be extremely verbose.

You can only stop the header from printing out (`verbose = false`) but not the footer. In addition I've moved the other failure messages to `$stderr`.

It's looking more and more like this logging functionality should be moved to another object for composability ease.