Only verify HTML fixtures on modern rubies.
This still gives us a high level of confidence in the HTML formatter,
while drastically reducing the complexity of the specs. We only require
a single fixture with this method.

I removed some of the textmate coverage, since it was incompatible with
this change and it's going away anyway so I don't care about it.

Backport from RSpec 3.