Removing broken test unnecessarilly scoped for Windows
Removing the scope for a test that does not need to be written
with different versions for Windows and Unix/Linux environments.
The implementation of ruby for windows (called mingw) makes the
behavior of a file_path be exactly the same on all platforms.

On MinGW, a File::SEPARATOR is '/' just like on *NIX like environment.
So it's pointless write tests with path separated as '/' to run just
on *NIX and tests with path separated by '\\' to run just on Windows