MemoizedHelpers can call super
This actually happened earlier, when I switched to Monitor
to simplify the implementation and pass a failing test,
but I made the spec fail a few different ways.

Strangely, I started getting "circular require" warnings,
which cause the spec to fail. I don't know why, but invoking
`eq` outside the threaded context seems to fix it, so there is
a before_all which does this. Would be nice to at least
understand why this happens.