allows rspec executable to take arguments from STDIN
This makes it possible to pipe arguments from the previous process's
STDOUT directly into rspec. Arguments are simply concated onto ARGV
and are handled by RSpec::Core::Runner as usual.

examples:
  `echo 'test_spec.rb' | rspec --color`
  `cat list_of_test_files.txt | rspec`
  `git status | grep -oe '\S*_spec.rb' | rspec`

This is ignored when STDIN is associated with a terminal device (tty),
i.e. when there is no process preceding and piped into rspec.