Don't clear out shared examples in clear_examples.
While technically examples, shared ones function more like
configuration. When using clear_examples, it is surprising that they are
removed since a shared library of them was likely loaded in the
initial configuration steps.

This is a change in behaviour. It was initially added in 2.9 in 8e8fb2b3af4422327e676d4e1c344268357f3536 with no rationale. Its API visibility was unspecified at the time, and it's been officially private since 3.

Reported (among other things) in #2367.