Add did_you_mean functionality to rspec-core.
The functionality provides for suggestions when a person enters an
rspec spec/some_file_pathcommand with  typographical errors.

If there are errors, then at lib/rspec/core/configuration.rb:2037
the code searches for suggestions. If useful suggestions are found
they are added to the reported exception at
lib/rspec/core/configuration.rb:2038.

The suggestions are done by the DidYouMean::Suggestions service
object, which calls the DidYouMean::Proximity service object.
These are found in lib/rspec/core/did_you_mean

The other change to the current code is to provide for autoloading
of the DidYouMean constants, this is done at lib/rspec/core.rb:142.

Two test files are added in the spec/rspec/core/did_you_mean directory.

This functionality was inspired by
https://github.com/yuki24/did_you_mean
and partly based on the jaro_winkler code from that source
https://github.com/yuki24/did_you_mean/blob/master
/lib/did_you_mean/jaro_winkler.rb. This code has an MIT licence.