only perform coverage checks on ruby 3.3
This allows us to check coverage on lines that only run on modern
rubies, rather than needing to :nocov: them so older rubies can pass the
test suite. It also makes it easy to keep track of and change which
ruby's coverage is being _used_ in CI, since it's just being specified
in an environment variable on the matrix.